+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add theme customizer
 - Add menus setup
 - Add metaboxes setup
+- Add post type templates setup
 
 ## [0.6.2] - 2020-02-08