+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add post type templates setup
 - Add scripts (JavaScript) setup
 - Add sidebars setup
+- Add styles setup
 
 ## [0.6.2] - 2020-02-08