+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add sidebars setup
 - Add styles setup
 - Add thumbnails setup
+- Add translations setup
 
 ## [0.6.2] - 2020-02-08